use vanilly process for nbminer
